requirements.txt
asgiref==3.3.4
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
docutils==0.17.1
feedgenerator==1.9.1
Flask==1.1.2
Frozen-Flask==0.15
ghp-import==1.1.0
idna==2.10
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
pelican==4.6.0
Pillow==8.2.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==1.0.1
